lucent-serverless-controller/pyproject.toml
[project]
name = "lucent-serverless-controller"
version = "0.1.0"
description = "Control plane for lucent-serverless: scheduler, reaper, and /resolve."
requires-python = ">=3.10"
dependencies = [
    "lucent-serverless",
    "fastapi>=0.115,<0.120",
    "uvicorn[standard]>=0.32,<0.35",
    "httpx>=0.27,<0.30",
    "python-dotenv>=1.0,<2.0",
]

[tool.uv.sources]
lucent-serverless = { workspace = true }

[project.scripts]
lucent-controller = "lucent_serverless_controller.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lucent_serverless_controller"]
